requirements.txt
pyserial>=3.5
sounddevice>=0.4.6
numpy>=1.21.0
soundfile>=0.12.1
scipy>=1.7.0
websockets>=10.0
python-dotenv>=0.19.0

